moved navigation to features
